requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
opencv-python==4.9.0.80
numpy==1.26.4
tensorflow==2.16.1
scikit-learn==1.5.0
matplotlib==3.9.0
Pillow==10.3.0
ultralytics==8.2.20
